refactor: restructure features
