+++ CoShift/frontend/src/features/auth/types/index.ts
@@ -1,0 +1,5 @@
+export interface AuthCtx {
+  header: string | null
+  balance: number | null
+  logout: () => void
+}
