+++ CoShift/frontend/src/features/week/types/index.ts
@@ -1,0 +1,8 @@
+export interface ShiftCellVM {
+  startTime: string
+  fullyStaffed: boolean
+}
+
+export interface DayCellViewModel {
+  shifts: ShiftCellVM[]
+}
