+++ CoShift/frontend/src/layout/PrivateLayout.tsx
@@ -1,3 +1,3 @@
 import { Navigate, Outlet } from 'react-router-dom'
-import { useAuth } from '../feature/auth/AuthContext'
+import { useAuth } from '../features/auth/hooks/AuthContext'
 export default function PrivateLayout(){ return useAuth().header ? <Outlet/> : <Navigate to="/" replace/> }
